added histogram toolbar
