+++ src/components/Histogram/HistogramComponent.scss
@@ -9,8 +9,9 @@
         flex: 3 3 auto;
         min-width: 0;
         padding-bottom: 5px;
+
         .histogram-plot {
-            height: 100%;
+            height: calc(100% - 35px);
             position: relative;
         }
     }
