+++ src/components/Histogram/HistogramToolbarComponent/HistogramToolbarComponent.scss
