+++ src/components/SpectralProfiler/SpectralProfilerToolbarComponent/SpectralProfilerToolbarComponent.scss
@@ -1,0 +1,14 @@
+.spectral-profiler-toolbar {
+    overflow: hidden;
+    display: flex;
+
+    .bp3-form-group {
+        margin-bottom: 5px;
+        margin-top: 5px;
+        margin-left: 15px;
+
+        &:first-of-type {
+            margin-left: 5px;
+        }
+    }
+}
